Added React-Router; and changes App.js to single-page application
Added FontAwesome; started working on VideosPage component; moved entire DashboardPage component
