+++ src/RoutePages/VideosPage.js
@@ -1,10 +1,11 @@
 import React from 'react'
+import { Container, Row, Col, Card } from "react-bootstrap"
 
 function VideosPage() {
     return (
-        <div>
-            
-        </div>
+        <Container>
+            <h1>Youtube</h1>
+        </Container>
     )
 }
